+++ exp/services/webauth/internal/serve/serve.go
@@ -56,7 +56,7 @@ func handler(opts Options) (http.Handler, error) {
 		return nil, errors.Wrap(err, "parsing JWT private key")
 	}
 
-	horizonTimeout := 1 * time.Minute
+	horizonTimeout := horizonclient.HorizonTimeout
 	httpClient := &http.Client{
 		Timeout: horizonTimeout,
 	}
@@ -64,7 +64,7 @@ func handler(opts Options) (http.Handler, error) {
 		HorizonURL: opts.HorizonURL,
 		HTTP:       httpClient,
 	}
-	horizonClient.SetHorizonTimeOut(uint(horizonTimeout / time.Second))
+	horizonClient.SetHorizonTimeout(horizonTimeout)
 
 	mux := supporthttp.NewAPIMux(opts.Logger)